requirements.txt
streamlit==1.24.1
ultralytics==8.0.118
opencv-python-headless==4.7.0.72
pillow==9.5.0
numpy==1.24.4
pandas==2.0.2